+++ storage/buffer/src/main/java/io/atomix/storage/buffer/ByteBufferBuffer.java
@@ -28,7 +28,7 @@ public abstract class ByteBufferBuffer extends AbstractBuffer {
     this.bytes = bytes;
   }
 
-  public ByteBufferBuffer(ByteBufferBytes bytes, long offset, long initialCapacity, long maxCapacity, ReferenceManager<Buffer> referenceManager) {
+  public ByteBufferBuffer(ByteBufferBytes bytes, int offset, int initialCapacity, int maxCapacity, ReferenceManager<Buffer> referenceManager) {
     super(bytes, offset, initialCapacity, maxCapacity, referenceManager);
     this.bytes = bytes;
   }
@@ -39,11 +39,11 @@ public abstract class ByteBufferBuffer extends AbstractBuffer {
   }
 
   @Override
-  protected void compact(long from, long to, long length) {
+  protected void compact(int from, int to, int length) {
     byte[] bytes = new byte[1024];
-    long position = from;
+    int position = from;
     while (position < from + length) {
-      long size = Math.min((from + length) - position, 1024);
+      int size = Math.min((from + length) - position, 1024);
       this.bytes.read(position, bytes, 0, size);
       this.bytes.write(0, bytes, 0, size);
       position += size;
